Added Comments and small refactorings
